+++ src/components/graph/index.ts
@@ -1,1 +1,2 @@
 export * from './graph'
+export * from './graph.context'
